feat(etl)!: honour every DuckDB scalar field type, fail closed on the rest
Operator ask: offer all DuckDB-supported data types in the schema grid.
That was an ENGINE change, not a dropdown change — the 4-entry UI list
was a deliberate honesty guard, and its own comment said so:
TransformCompiler.direct() TRY_CAST'd only TIMESTAMP/DATE/DOUBLE and its
`default` branch emitted the column UNCAST. So `type: BIGINT` silently
produced a VARCHAR store, with nothing validating the declared type.
Widening the list alone would have industrialised that silent loss.

New SchemaFieldTypes is the ONE authority on the honoured vocabulary and
the SQL each type compiles to, replacing three copies of the old
three-type list that were free to drift: the cast switch (direct), the
date-vs-else choice (dateExpr), and cast-failure measurability
(DataTransformer.coercedSourceColumn). That third site is the
load-bearing one — a newly honoured type would otherwise have cast, and
silently nulled values, WITHOUT the failure audit counting them.

Honoured: VARCHAR, BOOLEAN, TINYINT..HUGEINT and unsigned twins,
FLOAT/DOUBLE, DECIMAL(p,s), DATE, TIME, TIMESTAMP, TIMESTAMPTZ, UUID,
BLOB. Date-like types keep the strptime-format-list path (after the
load-bearing CAST AS VARCHAR); everything else is TRY_CAST, so a bad
value still nulls rather than failing the batch. Refused, fail closed at
config load (operator decision): nested LIST/STRUCT/MAP (a text token
cannot become one), JSON (extension-dependent, and we do not statically
link DuckDB extensions), INTERVAL. Blank/absent still means VARCHAR, so
existing schemas are untouched.

UI mirrors it: 20 types with per-type icons and plain-words hints, and
DECIMAL reveals precision/scale inputs clamped to 1-38 / 0-p so an
out-of-range value can never reach the new gate. narrowToSchemaType no
longer collapses BIGINT into DOUBLE, which was lossy above 2^53 —
exactly where long numeric identifiers live.

BREAKING CHANGE: a schema declaring a type the engine cannot cast is now
refused at config load instead of silently storing text, and a config
declaring INTEGER (three demo/template spaces) now produces a real
INTEGER — changing that store's output Parquet schema on the next write,
which can trip the schema BACKWARD gate.

Gates: Java 3507/0/0/5 across 19 modules exit 0 (TypeFlowTest's
footer-parity check included, so the widened casts agree with what the
written Parquet carries). UI lint:tokens 0, 3x tsc 0, test:ci 2691
passed / 5 skipped exit 0, AOT 0. Preview-verified: the full type menu
renders, DECIMAL authors DECIMAL(18,2) with per-row precision/scale
inputs that clamp 99 -> 38 and scale 50 -> 38, and auto-detect now seeds
BIGINT instead of DOUBLE.

+++ inspecto-ui/src/app/modules/admin/pipelines/pipeline-parse-definition.component.spec.ts
@@ -1194,8 +1194,9 @@ describe('PipelineParseDefinitionComponent', () => {
             p.onPreviewed(typedPreview);
             fixture.detectChanges();
 
-            // BIGINT narrows to DOUBLE — the grid's honest vocabulary.
-            expect(p.schemaSeed().map((r) => r.type)).toEqual(['DOUBLE', 'DATE', 'VARCHAR']);
+            // BIGINT is KEPT since 2026-08-22 — the engine casts every DuckDB scalar type, so
+            // collapsing it into DOUBLE (lossy above 2^53) is no longer honest, just lossy.
+            expect(p.schemaSeed().map((r) => r.type)).toEqual(['BIGINT', 'DATE', 'VARCHAR']);
         });
 
         it('seeds VARCHAR when the server serves no columnTypes (old server, byte-identical)', async () => {
@@ -1219,7 +1220,7 @@ describe('PipelineParseDefinitionComponent', () => {
             const raw = schemaWrites[0].config['raw'] as Record<string, unknown>;
             expect(raw['types']).toBe('auto');
             const fields = raw['fields'] as { name: string; type: string; synonym?: string }[];
-            expect(fields.map((f) => f.type)).toEqual(['DOUBLE', 'DATE', 'VARCHAR']);
+            expect(fields.map((f) => f.type)).toEqual(['BIGINT', 'DATE', 'VARCHAR']);
             expect('synonym' in fields[0]).toBe(false);
         });
 
@@ -1231,7 +1232,7 @@ describe('PipelineParseDefinitionComponent', () => {
             expect(p.schemaSeed().map((r) => r.type)).toEqual(['VARCHAR', 'VARCHAR', 'VARCHAR']);
 
             p.applySuggestedTypes();
-            expect(p.schemaSeed().map((r) => r.type)).toEqual(['DOUBLE', 'DATE', 'VARCHAR']);
+            expect(p.schemaSeed().map((r) => r.type)).toEqual(['BIGINT', 'DATE', 'VARCHAR']);
         });
 
         it('a hydrated schema without the raw.types marker loads as Declared', async () => {
